Changed prototype of printf_string_upper_bound to return gsize. Now
* glib/gmessages.c: Changed prototype of printf_string_upper_bound
to return gsize.  Now matches the actual function body.

* glib/gstrfuncs.c (g_strrstr_len): Changed type of parameter #2
from gint to gssize.  Now matches the prototype in gstrfuncs.h.

+++ glib/gmessages.c
@@ -71,7 +71,7 @@ struct _GLogHandler
 
 
 /* --- prototypes --- */
-static guint printf_string_upper_bound (const gchar *format,
+static gsize printf_string_upper_bound (const gchar *format,
 					gboolean     may_warn,
 					va_list      args);
